backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
mysql-connector-python==8.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
SpeechRecognition==3.10.1
pydantic==2.6.4
python-dotenv==1.0.1
